fix doctest
